API-Data-Engineer-master/requirements.txt
Flask==1.1.2
pandas==0.25.3
scikit-learn==0.23.1
google-cloud-storage==1.29.0
mlflow==1.7.2
dill==0.3.1.1
